\name{LPI-package}
\alias{LPI-package}
\docType{package}
\title{Lag Profile Inversion}
\description{ Incoherent scatter radar lag profile deconvolution by
  means of statistical inversion }
\details{
  \tabular{ll}{
    Package: \tab LPI \cr
    Version: \tab 0.4 \cr
    Date: \tab 2025-05-16\cr
    License: FreeBSD\tab \cr
    Depends: R ()\tab \cr
    Built: \tab \cr
  }
  Index:

  \code{\link{LPI}} Lag Profile Inversion


}

\author{
Ilkka Virtanen (University of Oulu, Finland) \cr \email{ilkka.i.virtanen@oulu.fi}
}

\keyword{ package }
