% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPI2Gamma.R
\name{LPI2Gamma_mle_helper}
\alias{LPI2Gamma_mle_helper}
\title{LPI2Gamma}
\usage{
LPI2Gamma_mle_helper(x)
}
\arguments{
\item{x}{random sample}

\item{L}{lower bound}

\item{p1}{first percentile}

\item{p3}{second percentile}

\item{R}{the number of boostrap sample}

\item{alpha}{confidence level
LPI2Gamma(x, L, p1, p3, R, alpha)}
}
\description{
LPI2Gamma
}
