% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPI2Lognormal.R
\name{LPI2LogNormal_mle_helper}
\alias{LPI2LogNormal_mle_helper}
\title{LPI2Lognormal}
\usage{
LPI2LogNormal_mle_helper(x)
}
\arguments{
\item{x}{random sample}

\item{L}{lower bound}

\item{p1}{first percentile}

\item{p3}{second percentile}

\item{R}{the number of boostrap sample}

\item{alpha}{confidence level
LPI2LogNormal(x, L, p1, p3, R, alpha)}
}
\value{
the output from list
}
\description{
LPI2Lognormal
}
