% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPI2Weibull.R
\name{LPI2Weibull}
\alias{LPI2Weibull}
\title{LPI2 Weibull Index}
\usage{
LPI2Weibull(x, L, p1 = 0.25, p3 = 0.75, R = 1000, alpha = 0.05)
}
\arguments{
\item{x}{numeric vector}

\item{L}{lower bound}

\item{p1}{first percentile}

\item{p3}{second percentile}

\item{R}{number of bootstrap replications}

\item{alpha}{significance level}
}
\value{
A data frame of estimation results
}
\description{
Computes the LPI2 index under the Weibull distribution.
}
