% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPI2Weibull.R
\name{LPI2Weibull_boot_worker}
\alias{LPI2Weibull_boot_worker}
\title{LPI2Weibull}
\usage{
LPI2Weibull_boot_worker(data, indices, p1, p3, L)
}
\arguments{
\item{p1}{first percentile}

\item{p3}{second percentile}

\item{L}{lower bound}

\item{x}{random sample}

\item{R}{the number of boostrap sample}

\item{alpha}{confidence level
LPI2LogNormal(x, L, p1, p3, R, alpha)}
}
\description{
LPI2Weibull
}
