% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPI3Gamma.R
\name{LPI3Gamma}
\alias{LPI3Gamma}
\title{LPI3 Gamma Index}
\usage{
LPI3Gamma(x, L, R = 1000, alpha = 0.05)
}
\arguments{
\item{x}{numeric vector}

\item{L}{lower bound}

\item{R}{number of bootstrap replications}

\item{alpha}{significance level}
}
\value{
data frame of results
}
\description{
Computes the LPI3 index under the Gamma distribution.
}
