% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPI3Gamma.R
\name{LPI3Gamma_mle_helper}
\alias{LPI3Gamma_mle_helper}
\title{LPI3Gamma}
\usage{
LPI3Gamma_mle_helper(x)
}
\arguments{
\item{x}{random sample}

\item{L}{lower bound}

\item{R}{the number of boostrap sample}

\item{alpha}{confidence level
LPI3Gamma(x, L, R, alpha)}
}
\description{
LPI3Gamma
}
