% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPI3Lognormal.R
\name{LPI3LogNormal_mle_helper}
\alias{LPI3LogNormal_mle_helper}
\title{LPI3Lognormal}
\usage{
LPI3LogNormal_mle_helper(x)
}
\arguments{
\item{x}{random sample}

\item{L}{lower bound}

\item{R}{the number of boostrap sample}

\item{alpha}{confidence level
LPI3LogNormal(x, L, R, alpha)}
}
\description{
LPI3Lognormal
}
