% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPI3Weibull.R
\name{LPI3Weibull}
\alias{LPI3Weibull}
\title{LPI3 Weibull Index}
\usage{
LPI3Weibull(x, L, R = 1000, alpha = 0.05)
}
\arguments{
\item{x}{numeric vector}

\item{L}{lower bound}

\item{R}{number of bootstrap replications}

\item{alpha}{significance level}
}
\value{
data frame of results
}
\description{
Computes the LPI3 index under the Weibull distribution.
}
