% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPI3Weibull.R
\name{LPI3Weibull_boot_worker}
\alias{LPI3Weibull_boot_worker}
\title{LPI3Weibull}
\usage{
LPI3Weibull_boot_worker(data, indices, L)
}
\arguments{
\item{L}{lower bound}

\item{x}{random sample}

\item{R}{the number of boostrap sample}

\item{alpha}{confidence level
LPI3Weibull(x, L, R, alpha)}
}
\description{
LPI3Weibull
}
