% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPIGamma.R
\name{LPIgamma}
\alias{LPIgamma}
\title{LPI Gamma Classic CL Index}
\usage{
LPIgamma(x, L, R = 1000, alpha = 0.05)
}
\arguments{
\item{x}{numeric vector}

\item{L}{lower bound}

\item{R}{number of bootstrap replications}

\item{alpha}{significance level}
}
\value{
data frame of results
}
\description{
Computes the classic capability index under the Gamma model.
}
