% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPIGamma.R
\name{LPIgamma_mle_helper}
\alias{LPIgamma_mle_helper}
\title{LPIGamma}
\usage{
LPIgamma_mle_helper(x)
}
\arguments{
\item{x}{random sample}

\item{L}{lower bound}

\item{R}{the number of boostrap sample
LPIgamma(x, L, R, alpha)}
}
\description{
LPIGamma
}
