% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPILognormal.R
\name{LPIlognormal_mle_helper}
\alias{LPIlognormal_mle_helper}
\title{Lognormal}
\usage{
LPIlognormal_mle_helper(x)
}
\arguments{
\item{x}{random sample}

\item{L}{lower bound}

\item{R}{the number of boostrap sample}

\item{alpha}{confidence level
LPIlognormal(x, L, R, alpha)}
}
\description{
Lognormal
}
