% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPIWeibull.R
\name{LPIWeibull}
\alias{LPIWeibull}
\title{LPI Weibull Classic CL Index}
\usage{
LPIWeibull(x, L, R = 1000, alpha = 0.05)
}
\arguments{
\item{x}{numeric vector}

\item{L}{lower bound}

\item{R}{number of bootstrap replications}

\item{alpha}{significance level}
}
\value{
data frame of results
}
\description{
Computes the classic capability index under the Weibull model.
}
