% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPIWeibull.R
\name{LPIWeibull_boot_worker}
\alias{LPIWeibull_boot_worker}
\title{LPIWeibull}
\usage{
LPIWeibull_boot_worker(data, indices, L)
}
\arguments{
\item{L}{lower bound}

\item{x}{random sample}

\item{R}{the number of boostrap sample}

\item{alpha}{confidence level
LPIWeibull(x, L, R, alpha)}
}
\description{
LPIWeibull
}
