\name{LPI-package}
\alias{LPI-package}
\docType{package}
\title{Lag Profile Inversion}
\description{ Lag Profile Inversion with support for computer clusters }
\details{
  \tabular{ll}{
    Package: \tab LPI \cr
    Version: \tab 0.3-5 \cr
    Date: \tab 2014-02-17\cr
    License: FreeBSD\tab \cr
    Depends: R ()\tab \cr
    Built: \tab \cr
  }
  Index:

  \code{\link{LPI}} Lag Profile Inversion


}

\author{
Ilkka Virtanen (University of Oulu, Finland) \cr \email{ilkka.i.virtanen@oulu.fi}
}

\keyword{ package }
